+++ morphman/common/vtk_wrapper.py
@@ -7,6 +7,92 @@ import vtk
 radiusArrayName = 'MaximumInscribedSphereRadius'
 
 
+def get_number_of_arrays(vtk_polydata):
+    """Returns the names and number of arrays for a vtkPolyData object
+
+    Args:
+        vtk_polydata (vtkPolyData): object / line to investigate the array.
+
+    Returns:
+        count (int): Number of arrays in the line.
+        names (list): A list of names of the arrays.
+    """
+    count = 0
+    names = []
+    name = 0
+    while name is not None:
+        name = vtk_polydata.GetPointData().GetArrayName(count)
+        if name is not None:
+            names.append(name)
+            count += 1
+
+    return count, names
+
+
+def extract_single_line(centerlines, line_id, start_id=0, end_id=None):
+    """Extract one line from multiple centerlines.
+    If startID and endID is set then only a segment of the centerline is extracted.
+
+    Args:
+        centerlines (vtkPolyData): Centerline to extract.
+        line_id (int): The line ID to extract.
+        start_id (int):
+        end_id (int):
+
+    Returns:
+        centerline (vtkPolyData): The single line extracted
+    """
+    cell = vtk.vtkGenericCell()
+    centerlines.GetCell(line_id, cell)
+    n = cell.GetNumberOfPoints() if end_id is None else end_id + 1
+
+    line = vtk.vtkPolyData()
+    cell_array = vtk.vtkCellArray()
+    cell_array.InsertNextCell(n - start_id)
+    line_points = vtk.vtkPoints()
+
+    arrays = []
+    n_, names = get_number_of_arrays(centerlines)
+    for i in range(n_):
+        tmp = centerlines.GetPointData().GetArray(names[i])
+        tmp_comp = tmp.GetNumberOfComponents()
+        radius_array = get_vtk_array(names[i], tmp_comp, n - start_id)
+        arrays.append(radius_array)
+
+    point_array = []
+    for i in range(n_):
+        point_array.append(centerlines.GetPointData().GetArray(names[i]))
+
+    count = 0
+    for i in range(start_id, n):
+        cell_array.InsertCellPoint(count)
+        line_points.InsertNextPoint(cell.GetPoints().GetPoint(i))
+
+        for j in range(n_):
+            num = point_array[j].GetNumberOfComponents()
+            if num == 1:
+                tmp = point_array[j].GetTuple1(i)
+                arrays[j].SetTuple1(count, tmp)
+            elif num == 2:
+                tmp = point_array[j].GetTuple2(i)
+                arrays[j].SetTuple2(count, tmp[0], tmp[1])
+            elif num == 3:
+                tmp = point_array[j].GetTuple3(i)
+                arrays[j].SetTuple3(count, tmp[0], tmp[1], tmp[2])
+            elif num == 9:
+                tmp = point_array[j].GetTuple9(i)
+                arrays[j].SetTuple9(count, tmp[0], tmp[1], tmp[2], tmp[3], tmp[4],
+                                    tmp[5], tmp[6], tmp[7], tmp[8])
+        count += 1
+
+    line.SetPoints(line_points)
+    line.SetLines(cell_array)
+    for j in range(n_):
+        line.GetPointData().AddArray(arrays[j])
+
+    return line
+
+
 def read_polydata(filename, datatype=None):
     """
     Load the given file, and return a vtkPolyData object for it.
