+++ morphman/manipulate_bend.py
@@ -8,7 +8,10 @@
 from argparse import ArgumentParser, RawDescriptionHelpFormatter
 
 # Local import
-from morphman.common import *
+from morphman.common.argparse_common import *
+from morphman.common.centerline_operations import *
+from morphman.common.surface_operations import *
+from morphman.common.voronoi_operations import *
 
 
 def manipulate_bend(input_filepath, output_filepath, smooth, smooth_factor, region_of_interest, region_points,
