+++ test/test_landmarking.py
@@ -9,7 +9,8 @@
 import pytest
 
 from .fixtures import common_input
-from morphman.common import get_path_names, extract_ica_centerline
+from morphman.common.common import get_path_names
+from morphman.common.centerline_operations import extract_ica_centerline
 from morphman.misc import landmarking_bogunovic, landmarking_piccinelli
