+++ test/test_manipulate_bifurcation.py
@@ -10,8 +10,9 @@ import pytest
 
 from .fixtures import common_input
 from morphman import manipulate_bifurcation
-from morphman.common import get_path_names, read_polydata, vtk_point_locator, get_centerline_tolerance, \
-    extract_single_line, get_distance
+from morphman.common.centerline_operations import get_centerline_tolerance
+from morphman.common.common import get_path_names, get_distance
+from morphman.common.vtk_wrapper import read_polydata, vtk_point_locator, extract_single_line
 
 
 @pytest.mark.parametrize("angle", [20 / 180 * np.pi, -20 / 180 * np.pi])
